requirements.txt
streamlit>=1.27.0
opencv-python-headless>=4.6.0.66
numpy>=1.22.0
streamlit-webrtc>=0.44.0
av>=10.0.0
aiortc>=1.5.0
openai>=1.0.0
python-dotenv>=1.0.1
Pillow>=9.0.0
requests>=2.28.0
pandas>=2.0.0